Update the `.oh-my-zsh` folder
The original .oh-my-zsh uploaded was a directory that contains a .git
folder, causing git to discard all the contents in it while staging
changes, because it was a git repository. This commit converts it to a
regular folder and backs all the files in it up on the Internet again.